Upgrade to version 2.0.0
This commit upgrades the `redux-action-analytics-middleware` package to
version `2.0.0`. Changes included in this commit are:

1) Updates to the README, related to the new features, and general
documentation additions
2) Additional rule to the `.eslintrc.json` file to require parameter
types. This is only relevant to developers of the project
3) Feature additions*

*The additional feature is an augmentation to the `trackCallback`
method. This callback method will now be passed the the value of the
state before and after the action is dispatched to allow for much richer
analytics (you may want to track the change to a value as a result of
dispatching an action, this was not possible in version `1.0.x`).